adding some changes to basically have my bar on my screen whenever i use my alias for hdmi lvds or dual

+++ .config/polybar/launch.sh
@@ -1,7 +1,25 @@
 #!/usr/bin/env bash
-killall -q polybar
 
-echo "---" | tee -a /tmp/polybarexample.log
-polybar example >>/tmp/polybarexample.log 2>&1 &
-
-echo "Bars launched..."
+if [ "$1" == "" ] ; then
+	echo "Missing parameter : hdmi , lvds, dual"
+else
+	killall -q polybar
+	
+	case $1 in
+	hdmi)
+		echo "---" | tee -a /tmp/polybarhdmi1.log
+		polybar hdmi1 >>/tmp/polybarhdm1.log 2>&1 &
+		;;
+	lvds)
+		echo "---" | tee -a /tmp/polybarlvds1.log
+		polybar lvds1 >>/tmp/polybarlvds1.log 2>&1 &
+		;;
+	dual)
+		echo "---" | tee -a /tmp/polybarlvds1.log /tmp/polybarhdmi1.log
+		polybar lvds1 >>/tmp/polybarlvds1.log 2>&1 &
+		polybar hdmi1 >>/tmp/polybarhdm1.log 2>&1 &
+		;;
+	esac
+	
+	echo "Bars launched..."
+fi
